+++ tests/unit/test_core_types.py
@@ -4,7 +4,6 @@ Testing that protocols work correctly and type aliases are valid.
 """
 
 from datetime import UTC, datetime
-from decimal import Decimal
 
 from src.core.enums import ActionType, PositionType, Symbol
 from src.core.models.position import Position, Trade
@@ -19,12 +18,12 @@ class TestProtocolCompliance:
         # Arrange
         position = Position(
             symbol=Symbol.BTC,
-            size=Decimal("1.0"),
-            entry_price=Decimal("50000.0"),
-            leverage=Decimal("10.0"),
+            size=float("1.0"),
+            entry_price=float("50000.0"),
+            leverage=float("10.0"),
             timestamp=datetime.now(UTC),
             position_type=PositionType.LONG,
-            margin_used=Decimal("5000.0"),
+            margin_used=float("5000.0"),
         )
 
         # Act & Assert - verify all protocol attributes exist
@@ -42,14 +41,14 @@ class TestProtocolCompliance:
         assert hasattr(position, "position_value")
 
         # Test method calls
-        pnl = position.unrealized_pnl(Decimal("52000.0"))
-        assert isinstance(pnl, Decimal)
+        pnl = position.unrealized_pnl(float("52000.0"))
+        assert isinstance(pnl, float)
 
-        risk = position.is_liquidation_risk(Decimal("45000.0"), Decimal("0.05"))
+        risk = position.is_liquidation_risk(float("45000.0"), float("0.05"))
         assert isinstance(risk, bool)
 
-        value = position.position_value(Decimal("52000.0"))
-        assert isinstance(value, Decimal)
+        value = position.position_value(float("52000.0"))
+        assert isinstance(value, float)
 
     def test_trade_implements_itrade_protocol(self) -> None:
         """Test that Trade class correctly implements ITrade protocol."""
@@ -88,21 +87,21 @@ class TestProtocolCompliance:
         # Arrange
         position = Position(
             symbol=Symbol.ETH,
-            size=Decimal("5.0"),
-            entry_price=Decimal("3000.0"),
-            leverage=Decimal("5.0"),
+            size=float("5.0"),
+            entry_price=float("3000.0"),
+            leverage=float("5.0"),
             timestamp=datetime.now(UTC),
             position_type=PositionType.SHORT,
-            margin_used=Decimal("3000.0"),
+            margin_used=float("3000.0"),
         )
 
         # Act - function that accepts IPosition protocol
-        def process_position(pos: IPosition) -> Decimal:
-            return pos.unrealized_pnl(Decimal("3100.0"))
+        def process_position(pos: IPosition) -> float:
+            return pos.unrealized_pnl(float("3100.0"))
 
         # Assert - Position should work as IPosition
         result = process_position(position)
-        assert isinstance(result, Decimal)
+        assert isinstance(result, float)
 
     def test_itrade_protocol_type_checking(self) -> None:
         """Test that ITrade protocol enables proper type checking."""
@@ -136,27 +135,27 @@ class TestTypeAliases:
         """Test PriceDict type alias works correctly."""
         # Arrange & Act
         prices: PriceDict = {
-            Symbol.BTC: Decimal("50000.0"),
-            Symbol.ETH: Decimal("3000.0"),
+            Symbol.BTC: float("50000.0"),
+            Symbol.ETH: float("3000.0"),
         }
 
         # Assert - should work as dict[Symbol, float]
         assert isinstance(prices, dict)
         assert Symbol.BTC in prices
-        assert prices[Symbol.BTC] == Decimal("50000.0")
-        assert prices[Symbol.ETH] == Decimal("3000.0")
+        assert prices[Symbol.BTC] == float("50000.0")
+        assert prices[Symbol.ETH] == float("3000.0")
 
     def test_position_dict_type_alias(self) -> None:
         """Test PositionDict type alias works correctly."""
         # Arrange
         btc_position = Position(
             symbol=Symbol.BTC,
-            size=Decimal("1.0"),
-            entry_price=Decimal("50000.0"),
-            leverage=Decimal("10.0"),
+            size=float("1.0"),
+            entry_price=float("50000.0"),
+            leverage=float("10.0"),
             timestamp=datetime.now(UTC),
             position_type=PositionType.LONG,
-            margin_used=Decimal("5000.0"),
+            margin_used=float("5000.0"),
         )
 
         eth_position = Position(
@@ -186,9 +185,9 @@ class TestTypeAliases:
         """Test that type aliases work in function signatures."""
 
         # Arrange
-        def calculate_portfolio_value(positions: PositionDict, prices: PriceDict) -> Decimal:
+        def calculate_portfolio_value(positions: PositionDict, prices: PriceDict) -> float:
             """Example function using type aliases."""
-            total_value = Decimal("0.0")
+            total_value = float("0.0")
             for symbol, position in positions.items():
                 if symbol in prices:
                     total_value += position.position_value(prices[symbol])
@@ -205,13 +204,13 @@ class TestTypeAliases:
         )
 
         positions: PositionDict = {Symbol.BTC: btc_position}
-        prices: PriceDict = {Symbol.BTC: Decimal("52000.0")}
+        prices: PriceDict = {Symbol.BTC: float("52000.0")}
 
         # Act
         result = calculate_portfolio_value(positions, prices)
 
         # Assert
-        assert isinstance(result, Decimal)
+        assert isinstance(result, float)
         assert result > 0
 
 
@@ -243,18 +242,14 @@ class TestProtocolContractVerification:
 
         # Act & Assert - unrealized_pnl contract
         long_pnl_profit = long_position.unrealized_pnl(
-            Decimal("52000.0")
+            float("52000.0")
         )  # Price up = profit for long
-        long_pnl_loss = long_position.unrealized_pnl(
-            Decimal("48000.0")
-        )  # Price down = loss for long
+        long_pnl_loss = long_position.unrealized_pnl(float("48000.0"))  # Price down = loss for long
 
         short_pnl_profit = short_position.unrealized_pnl(
-            Decimal("2800.0")
+            float("2800.0")
         )  # Price down = profit for short
-        short_pnl_loss = short_position.unrealized_pnl(
-            Decimal("3200.0")
-        )  # Price up = loss for short
+        short_pnl_loss = short_position.unrealized_pnl(float("3200.0"))  # Price up = loss for short
 
         assert long_pnl_profit > 0
         assert long_pnl_loss < 0
@@ -262,8 +257,8 @@ class TestProtocolContractVerification:
         assert short_pnl_loss < 0
 
         # position_value contract
-        long_value = long_position.position_value(Decimal("52000.0"))
-        short_value = short_position.position_value(Decimal("2800.0"))
+        long_value = long_position.position_value(float("52000.0"))
+        short_value = short_position.position_value(float("2800.0"))
 
         assert long_value > 0
         assert short_value > 0  # Value should always be positive (absolute position value)
